requirements.txt
anthropic
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
ddgs
python-dotenv
